+++ v3/internal/commands/appimage_testfiles/main.go
@@ -28,13 +28,13 @@ func main() {
 		log.Println("ApplicationDidFinishLaunching")
 	})
 
-	var hiddenWindows []*application.WebviewWindow
+	var hiddenWindows []application.Window
 
-	currentWindow := func(fn func(window *application.WebviewWindow)) {
+	currentWindow := func(fn func(window application.Window)) {
 		if app.Window.Current() != nil {
 			fn(app.Window.Current())
 		} else {
-			println("Current WebviewWindow is nil")
+			println("Current Window is nil")
 		}
 	}
 
@@ -60,7 +60,7 @@ func main() {
 	myMenu.Add("New WebviewWindow (Hides on Close one time)").
 		SetAccelerator("CmdOrCtrl+H").
 		OnClick(func(ctx *application.Context) {
-			w := app.Window.New()
+			w := application.Window(app.Window.New())
 			w.RegisterHook(events.Common.WindowClosing, func(e *application.WindowEvent) {
 				if !lo.Contains(hiddenWindows, w) {
 					hiddenWindows = append(hiddenWindows, w)
@@ -201,141 +201,141 @@ func main() {
 
 	sizeMenu := menu.AddSubmenu("Size")
 	sizeMenu.Add("Set Size (800,600)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetSize(800, 600)
 		})
 	})
 
 	sizeMenu.Add("Set Size (Random)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetSize(rand.Intn(800)+200, rand.Intn(600)+200)
 		})
 	})
 	sizeMenu.Add("Set Min Size (200,200)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetMinSize(200, 200)
 		})
 	})
 	sizeMenu.Add("Set Max Size (600,600)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetMaximiseButtonState(application.ButtonDisabled)
 			w.SetMaxSize(600, 600)
 		})
 	})
 	sizeMenu.Add("Get Current WebviewWindow Size").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			width, height := w.Size()
 			application.InfoDialog().SetTitle("Current WebviewWindow Size").SetMessage("Width: " + strconv.Itoa(width) + " Height: " + strconv.Itoa(height)).Show()
 		})
 	})
 
 	sizeMenu.Add("Reset Min Size").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetMinSize(0, 0)
 		})
 	})
 
 	sizeMenu.Add("Reset Max Size").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetMaxSize(0, 0)
 			w.SetMaximiseButtonState(application.ButtonEnabled)
 		})
 	})
 	positionMenu := menu.AddSubmenu("Position")
 	positionMenu.Add("Set Relative Position (0,0)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetRelativePosition(0, 0)
 		})
 	})
 	positionMenu.Add("Set Relative Position (Random)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetRelativePosition(rand.Intn(1000), rand.Intn(800))
 		})
 	})
 
 	positionMenu.Add("Get Relative Position").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			x, y := w.RelativePosition()
 			application.InfoDialog().SetTitle("Current WebviewWindow Position").SetMessage("X: " + strconv.Itoa(x) + " Y: " + strconv.Itoa(y)).Show()
 		})
 	})
 
 	positionMenu.Add("Set Position (0,0)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetPosition(0, 0)
 		})
 	})
 
 	positionMenu.Add("Set Position (Random)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetPosition(rand.Intn(1000), rand.Intn(800))
 		})
 	})
 
 	positionMenu.Add("Get Position").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			x, y := w.Position()
 			application.InfoDialog().SetTitle("Current WebviewWindow Position").SetMessage("X: " + strconv.Itoa(x) + " Y: " + strconv.Itoa(y)).Show()
 		})
 	})
 
 	positionMenu.Add("Center").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.Center()
 		})
 	})
 	stateMenu := menu.AddSubmenu("State")
 	stateMenu.Add("Minimise (for 2 secs)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.Minimise()
 			time.Sleep(2 * time.Second)
 			w.Restore()
 		})
 	})
 	stateMenu.Add("Maximise").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.Maximise()
 		})
 	})
 	stateMenu.Add("Fullscreen").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.Fullscreen()
 		})
 	})
 	stateMenu.Add("UnFullscreen").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.UnFullscreen()
 		})
 	})
 	stateMenu.Add("Restore").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.Restore()
 		})
 	})
 	stateMenu.Add("Hide (for 2 seconds)").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.Hide()
 			time.Sleep(2 * time.Second)
 			w.Show()
 		})
 	})
 	stateMenu.Add("Always on Top").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetAlwaysOnTop(true)
 		})
 	})
 	stateMenu.Add("Not always on Top").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetAlwaysOnTop(false)
 		})
 	})
 	stateMenu.Add("Google.com").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetURL("https://google.com")
 		})
 	})
 	stateMenu.Add("wails.io").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetURL("https://wails.io")
 		})
 	})
@@ -352,7 +352,7 @@ func main() {
 		}
 	})
 	stateMenu.Add("Get Screen for WebviewWindow").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			screen, err := w.GetScreen()
 			if err != nil {
 				application.ErrorDialog().SetTitle("Error").SetMessage(err.Error()).Show()
@@ -363,7 +363,7 @@ func main() {
 		})
 	})
 	stateMenu.Add("Disable for 5s").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			w.SetEnabled(false)
 			time.Sleep(5 * time.Second)
 			w.SetEnabled(true)
@@ -372,7 +372,7 @@ func main() {
 
 	if runtime.GOOS == "windows" {
 		stateMenu.Add("Flash Start").OnClick(func(ctx *application.Context) {
-			currentWindow(func(w *application.WebviewWindow) {
+			currentWindow(func(w application.Window) {
 				time.Sleep(2 * time.Second)
 				w.Flash(true)
 			})
@@ -381,7 +381,7 @@ func main() {
 
 	printMenu := menu.AddSubmenu("Print")
 	printMenu.Add("Print").OnClick(func(ctx *application.Context) {
-		currentWindow(func(w *application.WebviewWindow) {
+		currentWindow(func(w application.Window) {
 			_ = w.Print()
 		})
 	})