+++ v3/pkg/application/window_manager.go
@@ -70,7 +70,7 @@ func (wm *WindowManager) NewWithOptions(windowOptions WebviewWindowOptions) *Web
 }
 
 // Current returns the current active window (may be nil)
-func (wm *WindowManager) Current() *WebviewWindow {
+func (wm *WindowManager) Current() Window {
 	if wm.app.impl == nil {
 		return nil
 	}
@@ -78,10 +78,7 @@ func (wm *WindowManager) Current() *WebviewWindow {
 	wm.app.windowsLock.RLock()
 	defer wm.app.windowsLock.RUnlock()
 	result := wm.app.windows[id]
-	if result == nil {
-		return nil
-	}
-	return result.(*WebviewWindow)
+	return result
 }
 
 // Add adds a window to the manager